redirection divided in files and updated minishell.h
